test(keybindings): pin the window-shortcut layout path with a divergent layout
The QWERTY case asserted a code whose layout character equals its physical
fallback, so it passed whether or not resolveWindowShortcutAction consulted
the lookup. An AZERTY-shaped chord (physical KeyQ typing 'a') plus a negative
case for a physical KeyA that types another letter fail if the wiring is
dropped.

Moves the layout-resolution cases to their own file; adding them in place
carried window-shortcut-policy.test.ts past the max-lines cap.
